Detect new/rename columns from binlog message and update schema
